it works

+++ src/main/kotlin/Individual.kt
@@ -1,4 +1,4 @@
 import shapes.Shape
 import java.awt.Graphics
 
-data class Individual(val dna: DNA, val fitness: Double) {}
+data class Individual(val dna: DNA, val fitness: Int) {}
